tic tac toe
This project is an object-oriented program that was all put together and implement through a game of Tic-Tac-Toe. The goal is to illustrate the process of designing an object-oriented solution to a small problem by simulating Tic-Tac-Toe, between the human and the computer that lets them mark their chosen cells with their respective characters and whoever marks three in a row, wins.